made a new code for cards. cards are not build in cards.js and moved there to frontpage and projects with different rules
