+++ src/app/header/header.js
@@ -1,8 +1,9 @@
+
 import styles from "./header.module.css";
 import Image from "next/image";
 import Link from "next/link";
 
-export default function Header() {
+export default async function Header() {
   return (
     <>
       <div className={styles.header}>
