+++ src/app/main/main.module.css
@@ -1,141 +1,13 @@
-.projects {
-  /*margin-left: 3%;
-    margin-right: 3%;
-    margin-top: 2%;
-    margin-bottom: 2%;*/
-  padding: 1%;
-  /*box-shadow: -1px -1px 3px #888888;*/
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  gap: 1rem;
-}
-
-.projects h2 {
+.mainProjects {
   font-family: "Manrope", sans-serif;
   font-weight: 600;
   font-size: 1.5rem;
   color: black;
   text-transform: uppercase;
   margin-bottom: 1rem;
-  margin-top: 1rem;
+  margin-top: 5rem;
   grid-column: span 3;
   text-align: center;
   /*text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;*/
   letter-spacing: 0.5rem;
-}
-
-.project {
-  grid-column: span 1;
-  text-align: center;
-  margin-top: 3rem;
-}
-
-.project h3 {
-  margin-bottom: 2%;
-  text-align: center;
-  font-size: 1.8rem;
-  font-weight: 600;
-}
-
-.image {
-  max-width: 90%;
-  height: auto;
-  border: 1px solid #eeecec;
-  padding: 1%;
-  border-radius: 10px;
-  max-height: 300px;
-  object-fit: cover;
-}
-
-.project p {
-  padding-left: 10%;
-  padding-right: 10%;
-  text-align: justify;
-  margin-top: 1rem;
-}
-
-.codelanguages {
-  list-style: none;
-  margin-bottom: 2rem;
-  padding-left: 0;
-  margin-top: 1rem;
-}
-
-.codelanguages li {
-  font-size: 0.8rem;
-  display: inline;
-  padding: 0.5rem 0.5rem;
-
-  margin: 0 1rem;
-  background-color: rgba(13, 117, 13, 0.452);
-  border-radius: 5px;
-}
-
-.linkContainer {
-  margin-top: 1rem;
-}
-
-.projectLink {
-  padding: 1rem;
-  margin-top: 1rem;
-  text-decoration: none;
-}
-
-.loading {
-  height: 60vh;
-}
-
-@media only screen and (max-width: 1100px) {
-  .projects {
-    grid-template-columns: 1fr 1fr;
-  }
-  .projects h2 {
-    grid-column: span 2;
-  }
-}
-
-@media only screen and (max-width: 800px) {
-  .projects {
-    grid-template-columns: 1fr;
-    grid-gap: 0rem;
-  }
-
-  .projects h2 {
-    grid-column: 1;
-    font-size: 1.2rem;
-    letter-spacing: 0.4rem;
-  }
-
-  .project {
-    grid-column: 1;
-  }
-}
-
-@media only screen and (max-width: 500px) {
-  .projects h2 {
-    grid-column: 1;
-    font-size: 1.2rem;
-    letter-spacing: 0.3rem;
-  }
-
-  .projects p {
-    padding-left: 2%;
-    padding-right: 2%;
-  }
-
-  .project {
-    padding-left: 0%;
-    padding-right: 0%;
-  }
-
-  .projectLink {
-    padding: 0.5rem;
-  }
-
-  .codelanguages li {
-    font-size: 0.6rem;
-    display: inline;
-    padding: 0.2rem 0.2rem;
-    margin: 0.1rem;
-  }
 }
